--- src/shared/lib/form/index.ts
@@ -1,2 +1,0 @@
-export { default as useAppForm } from './model/hooks/useAppForm';
-export { default as useFieldValue } from './model/hooks/useFieldValue';
