+++ src/shared/lib/index.ts
@@ -1,1 +1,2 @@
 export { default as useGlobalErrorStore } from './error/useGlobalErrorStore';
+export { default as useAppForm } from './form/model/hooks/useAppForm';
