+++ src/shared/ui/form/NumberField.tsx
@@ -1,5 +1,5 @@
 import Input from '../Input';
-import { useFieldValue } from '../../lib/form';
+import useFieldValue from '../../lib/form/model/hooks/useFieldValue';
 import FormField from './FormField';
 
 interface NumberFieldProps {
