+++ src/shared/ui/form/SelectField.tsx
@@ -1,5 +1,5 @@
 import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from '../Select';
-import { useFieldValue } from '../../lib/form';
+import useFieldValue from '../../lib/form/model/hooks/useFieldValue';
 import FormField from './FormField';
 
 interface SelectOption {
