+++ src/shared/ui/form/TextField.tsx
@@ -1,5 +1,5 @@
 import { Input } from '@/shared/ui';
-import { useFieldValue } from '../../lib/form';
+import useFieldValue from '../../lib/form/model/hooks/useFieldValue';
 import FormField from './FormField';
 
 interface TextFieldProps {
